Move virtual connection stuff into HttpVCTable.h
This introduces HttpVCTable.h and HttpVCTable.cc to hold all the
stuff related to the HttpVCTable struct.
